refactor: redis to infra/redis
