+++ src/main/kotlin/com/relogging/server/infrastructure/admin/controller/AdminController.kt
@@ -8,7 +8,7 @@ import com.relogging.server.domain.plogging.dto.PloggingEventResponse
 import com.relogging.server.domain.plogging.service.PloggingEventService
 import com.relogging.server.domain.ploggingMeetup.dto.PloggingMeetupResponse
 import com.relogging.server.domain.ploggingMeetup.service.PloggingMeetupService
-import com.relogging.server.infrastructure.crawling.service.CrawlingService
+import com.relogging.server.infrastructure.scraping.service.ScrapingService
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.tags.Tag
 import jakarta.validation.Valid
@@ -29,9 +29,9 @@ import org.springframework.web.multipart.MultipartFile
 @Tag(name = "Admin", description = "admin 컨트롤러")
 class AdminController(
     @Qualifier("newsArticleCrawlingService")
-    private val newsArticleCrawlingService: CrawlingService,
+    private val newsArticleScrapingService: ScrapingService,
     @Qualifier("ploggingEventCrawlingService")
-    private val ploggingEventCrawlingService: CrawlingService,
+    private val ploggingEventScrapingService: ScrapingService,
     private val newsArticleService: NewsArticleService,
     private val ploggingEventService: PloggingEventService,
     private val ploggingMeetupService: PloggingMeetupService,
@@ -58,7 +58,7 @@ class AdminController(
     @Operation(summary = "뉴스 아티클 크롤링하기")
     @PostMapping("/newsArticles/crawl")
     fun startCrawling(): ResponseEntity<String> {
-        val count = newsArticleCrawlingService.crawlAndSave()
+        val count = newsArticleScrapingService.scrapNewsArticle()
         return ResponseEntity.ok("뉴스 아티클 $count 개 크롤링 성공했습니다")
     }
