+++ src/main/kotlin/com/relogging/server/security/oauth/handler/OAuthAuthenticationSuccessHandler.kt
@@ -1,7 +1,7 @@
 package com.relogging.server.security.oauth.handler
 
 import com.relogging.server.domain.user.entity.User
-import com.relogging.server.redis.service.RefreshTokenService
+import com.relogging.server.infrastructure.redis.service.RefreshTokenService
 import com.relogging.server.security.details.PrincipalDetails
 import com.relogging.server.security.jwt.provider.TokenProvider
 import jakarta.servlet.http.Cookie
